Task 0

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,12 +1,17 @@
 #include "lists.h"
+/**
+ * listint_len - Function that print the length of the node and also print
+ * @h: Input value
+ * Return: always true
+ */
 
 size_t listint_len(const listint_t *h)
 {
-	size_t count = 0;
+	int count = 0;
 
 	while (h != NULL)
 	{
-		printf("%d\n",h->n);
+		printf("%d\n", h->n);
 		count = count + 1;
 		h = h->next;
 	}
